+++ v0/destinations/rs/transform.js
@@ -1,4 +1,4 @@
-const { processWarehouseMessage } = require("../util");
+const { processWarehouseMessage } = require("../../../warehouse");
 
 function processSingleMessage(message, schemaVersion) {
   return processWarehouseMessage("rs", message, schemaVersion);
