+++ warehouse/v0/util.js
@@ -1,0 +1,69 @@
+const reservedANSIKeywordsMap = require("../config/ReservedKeywords.json");
+
+const toSnakeCase = str => {
+  if (!str) return "";
+  return String(str)
+    .replace(/^[^A-Za-z0-9]*|[^A-Za-z0-9]*$/g, "")
+    .replace(/([a-z])([A-Z])/g, (m, a, b) => a + "_" + b.toLowerCase())
+    .replace(/[^A-Za-z0-9]+|_+/g, "_")
+    .toLowerCase();
+};
+
+const toSafeDBString = str => {
+  if (parseInt(str[0]) > 0) str = `_${str}`;
+  str = str.replace(/[^a-zA-Z0-9_]+/g, "");
+  return str.substr(0, 127);
+};
+
+function safeTableName(provider, name = "") {
+  let tableName = name;
+  if (tableName === "") {
+    tableName = "STRINGEMPTY";
+  }
+  if (provider === "snowflake") {
+    tableName = tableName.toUpperCase();
+  }
+  if (provider === "postgres") {
+    tableName = tableName.toLowerCase();
+  }
+  if (
+    reservedANSIKeywordsMap[provider.toUpperCase()][tableName.toUpperCase()]
+  ) {
+    tableName = "_" + tableName;
+  }
+  return tableName;
+}
+
+function safeColumnName(provider, name = "") {
+  let columnName = name;
+  if (columnName === "") {
+    columnName = "STRINGEMPTY";
+  }
+  if (provider === "snowflake") {
+    columnName = columnName.toUpperCase();
+  }
+  if (provider === "postgres") {
+    columnName = columnName.toLowerCase();
+  }
+  if (
+    reservedANSIKeywordsMap[provider.toUpperCase()][columnName.toUpperCase()]
+  ) {
+    columnName = "_" + columnName;
+  }
+  return columnName;
+}
+
+function transformTableName(name = "") {
+  return toSnakeCase(name);
+}
+
+function transformColumnName(name = "") {
+  return toSafeDBString(name);
+}
+
+module.exports = {
+  safeColumnName,
+  safeTableName,
+  transformColumnName,
+  transformTableName
+};
